Moved trait, added some tests
