+++ src/ValueObjects/HasCollectionMethods.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Kadena\Traits;
+namespace Kadena\ValueObjects;
 
 trait HasCollectionMethods
 {
